+++ src/index.ts
@@ -7,7 +7,7 @@ import txpbv4 from "@kinecosystem/agora-api/node/transaction/v4/transaction_serv
 import { SystemInstruction, Transaction as SolanaTransaction } from "@solana/web3.js";
 
 import { Client } from "./client";
-import { errorsFromProto, TransactionErrors } from "./errors";
+import { errorsFromSolanaTx, errorsFromStellarTx, TransactionErrors } from "./errors";
 import { PrivateKey, PublicKey } from "./keys";
 import { Memo } from "./memo";
 import { MemoInstruction, MemoProgram } from "./solana/memo-program";
@@ -208,6 +208,17 @@ export interface Payment {
 
     invoice?: Invoice
     memo?: string
+
+    // dedupeId is a unique identifier used by the service to help prevent the 
+    // accidental submission of the same intended transaction twice. 
+    
+    // If dedupeId is set, the service will check to see if a transaction
+    // was previously submitted with the same dedupeId. If one is found,
+    // it will NOT submit the transaction again, and will return the status 
+    // of the previously submitted transaction.
+    // 
+    // Only available on Kin 4.
+    dedupeId?: Buffer
 }
 
 // ReadOnlyPayment represents a payment where the sender's
@@ -379,6 +390,9 @@ export function txDataFromProto(item: txpbv4.HistoryItem, state: txpbv4.GetTrans
                 stringMemo = memoParams.data;
             }
         }
+        if (item.getTransactionError()) {
+            data.errors = errorsFromSolanaTx(solanaTx, item.getTransactionError()!);
+        }
     }
     else if (item.getStellarTransaction()?.getEnvelopeXdr()) {
         const envelope = xdr.TransactionEnvelope.fromXDR(Buffer.from(item.getStellarTransaction()!.getEnvelopeXdr()));
@@ -388,6 +402,13 @@ export function txDataFromProto(item: txpbv4.HistoryItem, state: txpbv4.GetTrans
         } else if (envelope.v0().tx().memo().switch() === xdr.MemoType.memoText()) {
             stringMemo = envelope.v0().tx().memo().text().toString();
         }
+
+        if (item.getTransactionError()) {
+            data.errors = errorsFromStellarTx(envelope, item.getTransactionError()!);
+        }    
+    } else {
+        // This case *shouldn't* happen since either a solana or stellar should be set
+        throw new Error("invalid transaction");
     }
 
     const payments: ReadOnlyPayment[] = [];
@@ -407,14 +428,10 @@ export function txDataFromProto(item: txpbv4.HistoryItem, state: txpbv4.GetTrans
     });
     data.payments = payments;
 
-    if (item.getTransactionError()) {
-        data.errors = errorsFromProto(item.getTransactionError()!);
-    }
-
     return data;
 }
 
-// EarnBatch is a batch of earn payments to be sent.
+// EarnBatch is a batch of earn payments to be sent in a transaction.
 export interface EarnBatch {
     sender: PrivateKey
     channel?: PrivateKey
@@ -422,7 +439,19 @@ export interface EarnBatch {
 
     memo?: string
 
+    // The length of `earns` must be less than or equal to 15.
     earns: Earn[]
+
+    // dedupeId is a unique identifier used by the service to help prevent the 
+    // accidental submission of the same intended transaction twice. 
+    
+    // If dedupeId is set, the service will check to see if a transaction
+    // was previously submitted with the same dedupeId. If one is found,
+    // it will NOT submit the transaction again, and will return the status 
+    // of the previously submitted transaction.
+    // 
+    // Only available on Kin 4.
+    dedupeId?: Buffer
 }
 
 // Earn represents a earn payment in an earn batch.
@@ -433,19 +462,22 @@ export interface Earn {
 }
 
 // EarnBatchResult contains the results from an earn batch.
-export class EarnBatchResult {
-    succeeded: EarnResult[];
-    failed: EarnResult[];
+export interface EarnBatchResult {
+    txId: Buffer
 
-    constructor() {
-        this.succeeded = [];
-        this.failed = [];
-    }
+    // If TxError is defined, the transaction failed.
+    txError?: Error
+
+    // earnErrors contains any available earn-specific error
+    // information.
+    //
+    // earnErrors may or may not be set if TxError is set. 
+    earnErrors?: EarnError[]
 }
 
-// EarnResult contains the result of a submitted earn.
-export interface EarnResult {
-    txId?: Buffer
-    earn: Earn
-    error?: Error
+export interface EarnError {
+    // The error related to an earn.
+    error: Error
+    // The index of the earn that caused the
+    earnIndex: number
 }