+++ CHANGELOG.md
@@ -2,6 +2,27 @@
 
 ## Unreleased
 
+## [0.3.10](https://github.com/kinecosystem/kin-node/releases/tag/0.3.10)
+- Add `dedupeId` support on payments (`Client.submitPayment`) and earn batches (`Client.submitEarnBatch`)
+- `Client.submitEarnBatch` now supports submitting only a single transaction and up to 15 earns
+- `EarnBatchResult` is now an interface with `txId`, `txError` and `earnErrors`
+
+## [0.3.9](https://github.com/kinecosystem/kin-node/releases/tag/0.3.9)
+- Add `PaymentErrors` on `TransactionErrors`
+- Fix parsing transaction error in `Client.submitPayment`
+
+## [0.3.8](https://github.com/kinecosystem/kin-node/releases/tag/0.3.8)
+- Set operation errors on `TransactionErrors`
+
+## [0.3.7](https://github.com/kinecosystem/kin-node/releases/tag/0.3.7)
+- Fix client error handling
+
+## [0.3.6](https://github.com/kinecosystem/kin-node/releases/tag/0.3.6)
+- Add optional `accountResolution` parameter to `Client.getBalance`
+
+## [0.3.5](https://github.com/kinecosystem/kin-node/releases/tag/0.3.5)
+- Create new accounts with different token account address
+
 ## [0.3.4](https://github.com/kinecosystem/kin-node/releases/tag/0.3.4)
 - Do not reject Kin 4 payments with channel set
 - Check for duplicate signers for Stellar transactions
@@ -20,7 +41,7 @@
 - Rename `txHash` to `txId` in `Client.getTransaction`, `TransactionData` and `EarnResult`
 - Add `defaultCommitment` to `ClientConfig`
 - Add optional `commitment` parameter to `Client` methods (`createAccount`, `getBalance`, `getTransaction`, `submitPayment`, `submitEarnBatch`)
-- Add optional `subsidizer` parameter to `Client.create_account`, `Payment`, and `EarnBatch`
+- Add optional `subsidizer` parameter to `Client.createAcount`, `Payment`, and `EarnBatch`
 - Add optional `senderResolution` and `destinationResolution` parameters to `Client.submitPayment` and `Client.submitEarnBatch`
 - Mark `tx_hash` property in `Event` as deprecated.
 - Mark `SignTransactionRequest.txHash()` as deprecated in favour of `SignTransactionRequest.txId()`.